make give_help() DRY
