+++ src/dialogxml/dialogs/strdlog.hpp
@@ -87,4 +87,8 @@ void showWarning(std::string str1, std::string str2, cDialog* parent = nullptr);
 /// @param parent Optionally, a parent dialog.
 void showWarning(std::string str1, cDialog* parent = nullptr);
 
+// Show a help window with the given messages from rsrc/strings/help.txt:
+void give_help(short help1, short help2);
+void give_help(short help1, short help2, cDialog& parent);
+
 #endif
